Refactored obj processing into a helper file to better match api folder coding/organization style.
