+++ server/apis/routes/githubroutes.js
@@ -5,7 +5,7 @@ var GITHUB = require('../github.js');
 // These routes are relative to the mounted router. Therefore '/' here is actaully '/api/github'. 
 
 router.get('/repos', function(req, res) {
-  var username = req.user || 'makersquare'
+  var username = req.user || 'chppr'
 
   GITHUB.getUserRepos(username)
   .then(function(resp) {
